add-missing-tests

+++ src/main/java/com/groupe2/microservicedataobject/dataobject/aws/AwsDataObject.java
@@ -1,11 +1,14 @@
 package com.groupe2.microservicedataobject.dataobject.aws;
 
+import com.groupe2.microservicedataobject.dataobject.DataObjectAlreadyExists;
 import com.groupe2.microservicedataobject.dataobject.DataObjectNotFoundException;
 import com.groupe2.microservicedataobject.dataobject.IDataObject;
+import com.groupe2.microservicedataobject.dataobject.PathContainsOtherObjectsException;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.core.waiters.WaiterResponse;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.*;
+import software.amazon.awssdk.services.s3.paginators.ListObjectsV2Iterable;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 import software.amazon.awssdk.services.s3.presigner.model.GetObjectPresignRequest;
 import software.amazon.awssdk.services.s3.presigner.model.PresignedGetObjectRequest;
@@ -15,6 +18,8 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.time.Duration;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class AwsDataObject implements IDataObject {
     private static final AwsCloudClient AWS_CLIENT = AwsCloudClient.getInstance();
@@ -43,6 +48,10 @@ public class AwsDataObject implements IDataObject {
      * @param dataBytes the data to upload as bytes
      */
     public void upload(byte[] dataBytes) {
+        if (exists()){
+            throw new DataObjectAlreadyExists();
+        }
+
         if (!AwsBucketHelper.bucketExists(bucketName)) {
             AwsBucketHelper.createBucket(bucketName);
         }
@@ -84,21 +93,47 @@ public class AwsDataObject implements IDataObject {
         }
     }
 
+    public void delete(){
+        delete(false);
+    }
+
     /**
      * Deletes the current object
      * @throws DataObjectNotFoundException if the file does not exist
      */
-    public void delete() {
+    public void delete(boolean recursive) {
         if (!AwsBucketHelper.bucketExists(bucketName) || !exists()) {
             throw new DataObjectNotFoundException();
         }
         try (S3Client s3 = S3Client.builder().credentialsProvider(AWS_CLIENT.getCredentialsProvider()).region(AWS_CLIENT.getRegion()).build()) {
-            DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
-                    .bucket(bucketName)
-                    .key(path)
-                    .build();
+            if (!recursive) {
+                ListObjectsRequest listObjects = ListObjectsRequest
+                        .builder()
+                        .bucket(bucketName)
+                        .prefix(path)
+                        .build();
 
-            s3.deleteObject(deleteObjectRequest);
+                ListObjectsResponse res = s3.listObjects(listObjects);
+                if (res.contents().size() > 1){
+                    throw new PathContainsOtherObjectsException();
+                }
+
+                DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
+                        .bucket(bucketName)
+                        .key(path)
+                        .build();
+                s3.deleteObject(deleteObjectRequest);
+                return;
+            }
+
+            ListObjectsV2Request request = ListObjectsV2Request.builder().bucket(bucketName).prefix(path).build();
+            ListObjectsV2Iterable list = s3.listObjectsV2Paginator(request);
+            for (ListObjectsV2Response response : list) {
+                List<S3Object> objects = response.contents();
+                List<ObjectIdentifier> objectIdentifiers = objects.stream().map(o -> ObjectIdentifier.builder().key(o.key()).build()).collect(Collectors.toList());
+                DeleteObjectsRequest deleteObjectsRequest = DeleteObjectsRequest.builder().bucket(bucketName).delete(Delete.builder().objects(objectIdentifiers).build()).build();
+                s3.deleteObjects(deleteObjectsRequest);
+            }
         }
 
     }
@@ -139,14 +174,23 @@ public class AwsDataObject implements IDataObject {
      */
     public boolean exists() {
         try (S3Client s3 = S3Client.builder().credentialsProvider(AWS_CLIENT.getCredentialsProvider()).region(AWS_CLIENT.getRegion()).build()) {
+            if (!AwsBucketHelper.bucketExists(bucketName)){
+                return false;
+            }
 
-            GetObjectRequest request = GetObjectRequest.builder().bucket(bucketName).key(path).build();
-            s3.getObject(request);
+            ListObjectsRequest listObjects = ListObjectsRequest
+                    .builder()
+                    .bucket(bucketName)
+                    .prefix(path)
+                    .build();
+
+            ListObjectsResponse res = s3.listObjects(listObjects);
+            return res.contents().size() != 0;
 
         } catch (NoSuchKeyException e) {
             return false;
         }
         // TODO créer une exception custom pour chaque type d'erreur qui peut survenir.
-        return true;
+        //return true;
     }
 }
